+++ src/JestExt.ts
@@ -10,6 +10,7 @@ import * as status from './statusBar'
 import { TestReconciliationState } from './TestReconciliationState'
 import { pathToJestPackageJSON } from './helpers'
 import { readFileSync } from 'fs'
+import { execSync } from 'child_process'
 import { Coverage, showCoverageOverlay } from './Coverage'
 import { updateDiagnostics, resetDiagnostics, failedSuiteCount } from './diagnostics'
 import { DebugCodeLensProvider } from './DebugCodeLens'
@@ -411,17 +412,41 @@ export class JestExt {
   }
 
   private getJestVersion(version: (v: number) => void) {
+    // if all else fails, fallback to the last pre-20 release
+    version(this.getJestVersionViaPackageJson() || this.getJestVersionViaCommand() || 18)
+  }
+
+  private getJestVersionViaPackageJson() {
     const packageJSON = pathToJestPackageJSON(this.pluginSettings)
+
     if (packageJSON) {
       const contents = readFileSync(packageJSON, 'utf8')
       const packageMetadata = JSON.parse(contents)
+
       if (packageMetadata['version']) {
-        version(parseInt(packageMetadata['version']))
-        return
+        return parseInt(packageMetadata['version'])
       }
     }
-    // Fallback to last pre-20 release
-    version(18)
+
+    return null
+  }
+
+  private getJestVersionViaCommand() {
+    if (this.pluginSettings.pathToJest) {
+      const relativeJestCmd = this.pluginSettings.pathToJest.split(' ')[0]
+      const versionCmd = `${path.join(this.pluginSettings.rootPath, relativeJestCmd)} --version`
+
+      try {
+        const stdOut = execSync(versionCmd)
+          .toString()
+          .trim()
+        const ver = stdOut.slice(1) // stdOut will be something like 'v22.0.4', so trim the 'v'
+
+        return parseInt(ver)
+      } catch {}
+    }
+
+    return null
   }
 
   /**